+++ apps/desktop/package.json
@@ -41,7 +41,6 @@
 		"@types/diff-match-patch": "^1.0.36",
 		"@types/git-url-parse": "^9.0.3",
 		"@types/lscache": "^1.3.4",
-		"@types/marked": "^5.0.2",
 		"@types/node": "^22.3.0",
 		"@types/postcss-pxtorem": "^6.0.3",
 		"@vitest/ui": "^3.0.3",
@@ -58,7 +57,6 @@
 		"git-url-parse": "^14.0.0",
 		"jsdom": "^24.1.1",
 		"lscache": "^1.3.2",
-		"marked": "^10.0.0",
 		"mocha": "^10.7.0",
 		"nanoevents": "^9.0.0",
 		"postcss": "^8.4.38",
