+++ apps/desktop/tsconfig.json
@@ -36,8 +36,6 @@
 		"src/**/*.svelte",
 		"e2e/tests/**/*.js",
 		"e2e/tests/**/*.ts",
-		"e2e/tests/**/*.svelte",
-		"../../packages/shared/src/lib/codeHighlight.ts",
-		"../../packages/shared/src/lib/codeHighlight.ts"
+		"e2e/tests/**/*.svelte"
 	]
 }
