+++ packages/ui/package.json
@@ -63,7 +63,9 @@
 		"svelte-check": "catalog:svelte",
 		"vite": "catalog:",
 		"vitest": "catalog:",
-		"svelte-french-toast": "^1.2.0"
+		"svelte-french-toast": "^1.2.0",
+		"marked": "^10.0.0",
+		"@types/marked": "^5.0.2"
 	},
 	"peerDependencies": {
 		"svelte": "^4.0.0 || ^5.0.0-0"
