+++ packages/ui/src/lib/markdown/Markdown.svelte
@@ -1,7 +1,17 @@
 <script lang="ts">
-	import MarkdownContent from '$components/MarkdownContent.svelte';
-	import { options } from '$lib/utils/markdownRenderers';
+	import MarkdownContent from '$lib/markdown/MarkdownContent.svelte';
 	import { Lexer } from 'marked';
+
+	const options = {
+		async: false,
+		breaks: true,
+		gfm: true,
+		pedantic: false,
+		renderer: null,
+		silent: false,
+		tokenizer: null,
+		walkTokens: null
+	};
 
 	interface Props {
 		content: string | undefined;
